[DS-707] Style / readability
git-svn-id: http://scm.dspace.org/svn/repo/dspace/trunk@5546 9c30dcfa-912a-0410-8fc2-9e0234be79fd

+++ dspace-xmlui/dspace-xmlui-api/src/main/java/org/dspace/app/xmlui/cocoon/OpenSearchGenerator.java
@@ -159,7 +159,7 @@ public class OpenSearchGenerator extends AbstractGenerator
         if (this.validity == null)
         {
                 long expiry = System.currentTimeMillis() +
-                ConfigurationManager.getIntProperty("websvc.opensearch.validity") * 60 * 60 * 1000;
+                    ConfigurationManager.getLongProperty("websvc.opensearch.validity") * 60 * 60 * 1000;
                 this.validity = new ExpiresValidity(expiry);
         }
         return this.validity;
